requirements.txt
# Основные библиотеки для работы с Django
Django==5.1.3
django-allauth==65.2.0
django-filter==24.3
django-modeltranslation==0.19.12
djangorestframework==3.15.2

# Библиотеки для работы с задачами и очередями
celery==5.4.0
kombu==5.4.2
amqp==5.3.1
billiard==4.2.1

# Утилиты для тестирования и статической типизации
mypy-extensions==1.0.0
pytest==8.3.5

# Библиотеки для работы с данными и датами
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.2

# Внешние библиотеки для безопасности и сетевых запросов
requests==2.32.3
certifi==2024.12.14
idna==3.10
urllib3==2.3.0
six==1.17.0

# Инструменты для настройки и конфигурации
python-decouple==3.8
packaging==24.2
platformdirs==4.3.6
path==17.0.0
pathspec==0.12.1

# Библиотеки для работы с текстом
Markdown==3.7
sqlparse==0.5.2

# Инструменты для работы с командной строкой и интерфейсами
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
prompt_toolkit==3.0.48
colorama==0.4.6
wcwidth==0.2.13

# Утилиты для форматирования и кода
black==24.10.0

# Хуки
pre-commit==4.2.0
